refactor(batch-exports): extract per-destination config into a registry
The batch export edit form mixed every destination's behaviour into two
large files via per-destination branch logic: defaults, required fields,
validation, payload serialize/deserialize, event-table schema, and ~800
lines of conditional JSX all lived behind `if (service === ...)` chains.

Replace this with a declarative `DestinationDefinition` registry under
`destinations/`, one file per destination (S3, Postgres, Redshift,
Snowflake, BigQuery, HTTP, Databricks, AzureBlob). Each definition owns
its defaults, requiredFields, optional validate/serialize/deserialize,
event-table overrides, integration usage, and its `Fields` component.
`batchExportConfigFormLogic` and the edit form are now destination-agnostic
and route everything through `DESTINATIONS[type]`.

The form's errors merge only applies a destination validator's result when
it has a message, so a valid (undefined) result can't clobber a "required"
error — preserving the pre-refactor behaviour. The regression suite from
the base PRs passes unchanged, proving behaviour is preserved.

Generated-By: PostHog Code
Task-Id: 20b58ffb-0051-442d-ab87-58a15613645b
